+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Issue #530: New checkbox for the print module to disable the auto adjustment of the print mask scale
 - Adds a bitbucket pipeline to run `npm audit` automatically.
 - Module `filterGeneral`: wfs filter via server (aka `external wfs filtering`) added
+- Added a `showGeographicIdentifier` attribute to be able to use the `geographicIdentifier` to display the gazetteer search result.
 
 ### Changed
 - Issue #674, #676: addWMS tool and orientation control are now using the CRS of the map instead of hard coded CRS